--- src/main/java/com/example/game_application_server/domain/entiry/Deamon.java
@@ -1,4 +1,0 @@
-package com.example.game_application_server.domain.entiry;
-
-public class Deamon {
-}
