--- src/main/java/com/example/game_application_server/domain/entiry/Position.java
@@ -1,4 +1,0 @@
-package com.example.game_application_server.domain.entiry;
-
-public class Position {
-}
