+++ src/main/java/com/example/game_application_server/domain/entity/Position.java
@@ -1,0 +1,4 @@
+package com.example.game_application_server.domain.entity;
+
+public class Position {
+}
