+++ src/test/java/com/example/game_application_server/domain/entity/PlayerTest.java
@@ -1,0 +1,2 @@
+package com.example.game_application_server.domain.entity;public class PlayerTest {
+}
